+++ src/AbstractAdapter.ts
@@ -11,10 +11,24 @@ import {
   ResultObjectNumber,
   ResultObjectStream,
 } from "./types";
+import { parseUrl } from "./util";
 
 export abstract class AbstractAdapter implements IStorage {
   protected _type = "abstract-adapter";
-  protected _config: AdapterConfig;
+  protected _config: AdapterConfig | null;
+  protected _configError: string | null = null;
+
+  constructor(config?: string | AdapterConfig) {
+    if (typeof config === "string") {
+      const { value, error } = parseUrl(config);
+      if (error) {
+        this._configError = error;
+      }
+      this._config = value;
+    } else {
+      this._config = { ...config };
+    }
+  }
 
   get type(): string {
     return this._type;
@@ -24,8 +38,16 @@ export abstract class AbstractAdapter implements IStorage {
     return this._config;
   }
 
+  get configError(): string {
+    return this._configError;
+  }
+
   getType(): string {
     return this.type;
+  }
+
+  getConfigError(): string {
+    return this.configError;
   }
 
   getConfiguration(): AdapterConfig {