+++ src/types.ts
@@ -44,6 +44,10 @@ export interface IStorage {
    */
   config: AdapterConfig;
 
+  getConfigError(): string;
+
+  configError: string;
+
   /**
    * Returns an object that contains both the options passed with the configuration and the
    * default options of the storage type if not overruled by the options you passed in.
@@ -200,28 +204,9 @@ export enum StorageType {
   MINIO = "minio",
 }
 
-export type JSON = {
-  [id: string]:
-    | string
-    | number
-    | boolean
-    | string[]
-    | number[]
-    | boolean[]
-    | { [id: string]: JSON };
-};
-
-// export interface Options {
-//   [key: string]: string | number | boolean;
-// }
-
-export type GenericKey = number | string | boolean | number[] | string[] | boolean[];
-
 export interface AdapterConfig {
-  // type: StorageType;
   type: string;
-  // [id: string]: GenericKey;
-  [id: string]: number | string | boolean | number[] | string[] | boolean[];
+  [id: string]: number | string | boolean | number[] | string[] | boolean[] | object;
 }
 
 export type BackblazeAxiosResponse = {
@@ -287,14 +272,7 @@ export enum S3Compatible {
 
 export type ParseUrlResult = {
   error: string | null;
-  value: {
-    type: string;
-    part1: string;
-    part2: string;
-    part3: string;
-    bucketName: string;
-    queryString: { [key: string]: string };
-  };
+  value: AdapterConfig;
 };
 
 export interface ResultObject {